Fix: merge standalone req and res stream e2e tests into one
